requirements.txt
django-ninja
ensure-file
requests
django
json5
cairosvg
